disable powder snow freezing and all effects
